fixed links

+++ backend/matrigyan/views.py
@@ -643,7 +643,7 @@ def createQuiz(request):
 def editQuiz(request, id):
 	print(id)
 	# return Response("reached")
-	quiz=Quiz.objects.get(id=id)
+	quiz=Quiz.objects.filter(id=id).first()
 	if not quiz:
 		return Response({"success":False,"message":"No quiz exists with id"+id})
 	# print(request.data)
@@ -671,7 +671,9 @@ def editQuiz(request, id):
 
 @api_view(['DELETE'])
 def deleteQuiz(request, id):
-	quiz = Quiz.objects.get(id=id)
+	quiz = Quiz.objects.filter(id=id).first()
+	if quiz is None:
+		return Response({"success":False,"message":"No quiz found."})
 	quiz.delete()
 	return Response({"success":True,"message":'Quiz deleted!'})
 
@@ -787,7 +789,9 @@ def editSolution(request, id):
 	
 @api_view(['GET'])
 def getQuestions(request, id):
-	quiz = Quiz.objects.get(id=id)
+	quiz = Quiz.objects.filter(id=id).first()
+	if quiz is None:
+		return Response({"success":False,"message":"No quiz found."})
 	questions = quiz.questions.all()
 	if questions is None:
 		return Response({"success":False,"message":"No questions."})
@@ -796,7 +800,9 @@ def getQuestions(request, id):
 
 @api_view(['DELETE'])
 def deleteQuestion(request, id):
-	question = Question.objects.get(id=id)
+	question = Question.objects.filter(id=id).first()
+	if question is None:
+		return Response({"success":False,"message":"No question."})
 	question.delete()
 	return Response({"success":True,"message":"Question deleted!"})
 
@@ -823,7 +829,9 @@ def addSolution(request, id):
 
 @api_view(['DELETE'])
 def deleteSolution(request, id):
-	solution = Solution.objects.get(id=id)
+	solution = Solution.objects.filter(id=id).first()
+	if solution is None:
+		return Response({"success":False,"message":"No solution."})
 	solution.delete()
 	return Response({"success":True, "message":"Solution deleted"})
 
@@ -832,8 +840,12 @@ def addOption(request, id):
 	option = OptionSerializer(data=request.data)
 	if option.is_valid():
 		option.save()
-		question = Question.objects.get(id=id)
-		option_object = Option.objects.get(id=option.data['id'])
+		question = Question.objects.filter(id=id).first()
+		if question is None:
+			return Response({"success":False,"message":"No question."})
+		option_object = Option.objects.filter(id=option.data['id']).first()
+		if option_object is None:
+			return Response({"success":False,"message":"No options."})
 		question.options.add(option_object)
 		question=QuestionSerializer(question)
 		return Response({"question":question.data,"option_added":option_object.value,"success":True})
@@ -841,13 +853,17 @@ def addOption(request, id):
 
 @api_view(['DELETE'])
 def deleteOption(request, id):
-	option = Option.objects.get(id=id)
+	option = Option.objects.filter(id=id).first()
+	if option is None:
+		return Response({"success":False,"message":"No option."})
 	option.delete()
 	return Response({"success":True,"message":"Option deleted"})
 
 @api_view(['GET'])
 def getOptions(request, id):
-	question = Question.objects.get(id=id)
+	question = Question.objects.filter(id=id).first()
+	if question is None:
+		return Response({"success":False,"message":"No question."})
 	options = question.options.all()
 	if options is None:
 		return Response({"success":False,"message":"No options."})
@@ -885,7 +901,9 @@ def getEvents(request,id):
 	
 @api_view(['DELETE'])
 def deleteEvent(request, id):
-	event = Event.objects.get(id=id)
+	event = Event.objects.filter(id=id).first()
+	if event is None:
+		return Response({"success":False,"message":"No event."})
 	event.delete()
 	return Response({"success":True, "message":"Event deleted."})
 # @csrf_exempt
@@ -944,6 +962,8 @@ def getQuizResponse(request,quiz_id):
 @api_view(['GET'])
 def getResources(request):
 	tags = ResourceTag.objects.all()
+	if tags is None:
+		return Response({"success":False,"message":"No resources."})
 	sections = []
 	for tag in tags:
 		resources = Resource.objects.filter(tagname__id=tag.id)
@@ -960,7 +980,9 @@ def getResources(request):
 
 @api_view(['GET'])
 def getParticularResource(request, id):
-	resource = Resource.objects.get(id=id)
+	resource = Resource.objects.filter(id=id).first()
+	if resource is None:
+		return Response({"success":False,"message":"No resources."})
 	serializedresource = ResourceSerializer(resource, many=False)
 	return Response({"success":True,"resource":serializedresource.data})
 
@@ -982,7 +1004,9 @@ def addResource(request, id):
 		new_tag = ResourceTag(name=tag_name)
 		new_tag.save()
 		resource = Resource(image=data['image'],description=data['description'],title=data['title'],file_url=data['file_url'])
-		creator = Educator.objects.get(id=id)
+		creator = Educator.objects.filter(id=id).first()
+		if creator is None:
+			return Response({"success":False,"message":"No educator."})
 		resource.creator = creator
 		resource.tagname = new_tag
 		resource.save()
@@ -991,7 +1015,9 @@ def addResource(request, id):
 		return Response({"success":True, "resource":ser_res.data,"tag":ser_tag.data})
 	old_tag = ResourceTag.objects.filter(name=tag_name)[0]
 	resource = Resource(image=data['image'],description=data['description'],title=data['title'],file_url=data['file_url'])
-	creator = Educator.objects.get(id=id)
+	creator = Educator.objects.filter(id=id).first()
+	if creator is None:
+		return Response({"success":False,"message":"No educator."})
 	resource.creator = creator
 	resource.tagname = old_tag
 	resource.save()
@@ -1002,27 +1028,35 @@ def addResource(request, id):
 
 @api_view(['DELETE'])
 def deleteResource(request, id):
-	resource = Resource.objects.get(id=id)
+	resource = Resource.objects.filter(id=id).first()
+	if resource is None:
+		return Response({"success":False,"message":"No resource."})
 	resource.delete()
 	return Response({"success":True, "message":"Resource deleted!"})
 
 @api_view(['GET'])
 def getUser(request,id):
-		user = User.objects.get(id=id)
+		user = User.objects.filter(id=id).first()
+		if user is None:
+			return Response({"success":False,"message":"No user."})
 		print(user)
 		print(user.id)
 		ser_user = UserSerializer(user, many=False)
 		stu = Student.objects.filter(user__id=id).first()
 		# print(stu.DoesNotExist)
 		if stu is None:
-			educator = Educator.objects.get(user__id=id)
+			educator = Educator.objects.filter(user__id=id).first()
+			if educator is None:
+				return Response({"success":False,"message":"No educator found."})
 			ser_educator = EducatorSerializer(educator, many=False)
 			return Response({"success":True, "is_student":False, "user":ser_user.data,"educator":ser_educator.data})
 			# student = Student.objects.get(user__id=user.id)
 			# ser_student = StudentSerializer(student, many=False)
 			# return Response({"success":True, "user":ser_user,"is_student":True,"student":ser_student})
 		else:
-			student = Student.objects.get(user__id=user.id)
+			student = Student.objects.filter(user__id=user.id).first()
+			if student is None:
+				return Response({"success":False,"message":"No student."})
 			ser_student = StudentSerializer(student, many=False)
 			return Response({"success":True, "user":ser_user.data,"is_student":True,"student":ser_student.data})
 			# educator = Educator.objects.get(id=user.id)
@@ -1041,7 +1075,9 @@ def getEducators(request):
 
 @api_view(['GET'])
 def getEducator(request, id):
-	educator = Educator.objects.get(id=id)
+	educator = Educator.objects.filter(id=id).first()
+	if educator is None:
+		return Response({"success":False,"message":"No educator."})
 	ser_educator = EducatorSerializer(educator, many=False)
 	return Response({"success":True, "educator":ser_educator.data})
 
@@ -1053,8 +1089,12 @@ def editStudent(request,id):
 	email = data['email']
 	username = data['username']
 	phone = data['phone']
-	user = User.objects.get(id=id)
-	student = Student.objects.get(user__id=id)
+	user = User.objects.filter(id=id).first()
+	if user is None:
+		return Response({"success":False,"message":"No user."})
+	student = Student.objects.filter(user__id=id).first()
+	if student is None:
+		return Response({"success":False,"message":"No student."})
 	user.first_name = first_name
 	user.last_name = last_name
 	user.email = email
@@ -1073,8 +1113,12 @@ def editEducator(request,id):
 	last_name = data['last_name']
 	email = data['email']
 	username = data['username']
-	user = User.objects.get(id=id)
-	educator = Educator.objects.get(user__id=id)
+	user = User.objects.filter(id=id).first()
+	if user is None:
+		return Response({"success":False,"message":"No user."})
+	educator = Educator.objects.filter(user__id=id).first()
+	if educator is None:
+		return Response({"success":False,"message":"No educator."})
 	user.first_name = first_name
 	user.last_name = last_name
 	user.email = email
@@ -1091,7 +1135,6 @@ def searchCourses(request, search):
 	print(search)
 	search = search.lower()
 	courses = Course.objects.all()
-	print(courses)
 	if courses is None:
 		return Response({"success":False,"message":"No courses."})
 	course_list = []
@@ -1110,7 +1153,6 @@ def searchCourses(request, search):
 @api_view(['GET'])
 def filterCategory(request, category):
 	courses = Course.objects.filter(category__category=category.lower())
-	print(courses)
 	if courses is None:
 		return Response({"success":False,"message":"No courses."})
 	else:
